Update project structure, introduce base classes
