+++ SECEdgar/filings/__init__.py
@@ -1,0 +1,3 @@
+from SECEdgar.base import _EDGARBase
+from SECEdgar.filings.base import (Filing10K, Filing10Q, Filing13F,
+                                   Filing4, Filing8K, FilingSD)
