+++ SECEdgar/utils/exceptions.py
